move exportCharacter to meta service
