+++ src/app/pages/settings/settings.page.ts
@@ -33,25 +33,7 @@ export class SettingsPage implements OnInit {
     this.setOption(GameOption.TickTimer, $event.detail.value);
   }
 
-
   exportCharacter(slot: number) {
-    this.store.selectOnce(data => data).subscribe(data => {
-      const ignoredKeys: string[] = [];
-
-      const charData = data.charselect.characters[slot];
-      const charName = charData.name;
-
-      const saveData = Object.keys(data).filter(key => !ignoredKeys.includes(key)).reduce((acc, key) => {
-        acc[key] = data[key];
-        return acc;
-      }, {} as any);
-
-      const fileName = `${charName}.qivan`;
-      const dataStr = 'data:text/json;charset=utf-8,' + encodeURIComponent(JSON.stringify(saveData));
-      const downloadAnchorNode = document.createElement('a');
-      downloadAnchorNode.setAttribute('href',     dataStr);
-      downloadAnchorNode.setAttribute('download', fileName);
-      downloadAnchorNode.click();
-    });
+    this.metaService.exportCharacter(this.store, slot);
   }
 }
